2_human in the Loop/requirements.txt
fastapi
uvicorn
anthropic
google-genai
requests
python-dotenv
